% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation_validation.R
\name{CArima}
\alias{CArima}
\title{Fit ARIMA model to univariate time series}
\usage{
CArima(
y,
order = c(0,0,0),
seasonal = list(order = c(0,0,0), period = 1),
plot = T
xreg = NULL,
include.constant = TRUE,
lambda = model$lambda,
biasadj = FALSE,
method = c("CSS-ML", "ML", "CSS"),
model = NULL,
x = y,
...
)
}
\arguments{
\item{y}{a univariate time series of class \code{ts} or a matrix with the values of the time series.}

\item{order}{a specification of the non-seasonal part of the ARIMA model: the three integer components \code{(p, d, q)} are the AR order, the order of differencing, and the MA order.}

\item{seasonal}{a specification of the seasonal part of the ARIMA model, plus the period (which defaults to frequency(x) if the \code{y} parameter is a \code{ts} object). This may be a list with components order and period, or just a numeric vector of length 3 which specifies the seasonal order. In the latter case the default period is used.}

\item{plot}{if \code{TRUE}, it will show all the plots in execution.}

\item{include.constant}{for undifferenced series it fits the mean of the time series, for differenced series it fits the mean of the differenced time series. Note that if there is more than one difference taken, no constant is included regardless of the value of this argument.}

\item{...}{additional arguments to be passed to \code{\link[forecast]{Arima}}.}
}
\value{
See the \code{\link[forecast]{Arima}} function in the forecast package. The additional objects returned are: 
   \item{period}{The time series period of the seasonal part.}
   \item{SBC}{The SBC value corresponding to the log-likelihood.}
   \item{cor.coef}{Correlation matrix of the parameters of the model.}
   \item{resid.acf}{ACF of the residuals obtained with the \code{\link[forecast]{Acf}} function of the forecast package.}
   \item{resid.pacf}{PACF of the residuals obtained with the \code{\link[forecast]{Pacf}} function of the forecast package.}
   \item{lbtests.df}{Table of the Ljung-Box tests for some lags of the residuals, obtained with the \code{\link[TSA]{LB.test}} function of the TSA package.}
   \item{lbtests.plot}{Barplot of the Ljung-Box pvalues of the residuals.}
   \item{residRM}{Range-mean plot of the residuals obtained with the \code{\link{rmplot}} function.}
   \item{residQQp}{Qqplot of the residuals.}
   \item{residDensity}{Plot of the residual density compared with the normal density.}
   \item{residMuTest}{T-test for the residual's mean, obtained with \code{\link[stats]{t.test}} function of the stats package.}
   \item{residShapiro}{Shapiro-Wilk normality test for the residuals, obtained with \code{\link[stats]{shapiro.test}} function of the stats package.}
}
\description{
Largely a wrapper for the \code{\link[forecast]{Arima}} function in the forecast package. It adds some usefull statistics and tests to validate and compare models.
}
\details{
See the \code{\link[forecast]{Arima}} function in the forecast package.
}
\examples{
#Some examples with AirPassengers
model1 <- CArima(as.vector(log(AirPassengers)), order = c(0, 1, 1), 
                 seasonal = list(order=c(0,0,1), period = 12))
model1 
summary(model1, plot=F)

model2 <- CArima(as.vector(log(AirPassengers)), order = c(0, 1, 1), 
                 seasonal = list(order=c(0,1,1), period = 12))
model2 
summary(model2, plot=F) 
}
\seealso{
\code{\link[lmtest]{coeftest}}, \code{\link[gridExtra]{grid.arrange}}
}
