% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_forecast.R
\name{compareARIMA}
\alias{compareARIMA}
\title{Compare ARIMA Models}
\usage{
compareARIMA(object, dates, m1, m2, m3, h = 10, plot = T)
}
\arguments{
\item{object}{a time series object or a matrix containing the time series values.}

\item{dates}{a vector or an array with the dates as \code{character} of each observation of the time series. This vector can be generated using the function \code{\link{dateSeq}}.}

\item{m1}{an object describing a time series model; e.g. of class \code{\link{CArima}}.}

\item{m2}{an object describing a time series model; e.g. of class \code{\link{CArima}}.}

\item{m3}{an object describing a time series model; e.g. of class \code{\link{CArima}}.}

\item{h}{number of points that will not be used to fit the model, but to use as testing. The points will be taken from the bottom.}

\item{plot}{if \code{TRUE}, it will show all the plots in execution.}
}
\value{
An object of class \code{compareARIMA} containing:
   \item{metrics}{Some metrics of the models, such as AIC, Sigma square, SBC, SSE of prediction and two Ljung-Box test values.}
   \item{SSEtable}{A merge of the \code{SSEtable} of m1 and m2 obtained with \code{\link{predCap}} function.}
   \item{forecasts}{A merge of the \code{forecasts} of m1 and m2 obtained with \code{\link{predCap}} function.}
   \item{forecPlot1}{A plot of the forecasts values and the confidence intervals calculated of m1 obtained with \code{\link{predCap}} function.}
   \item{forecPlot2}{A plot of the forecasts values and the confidence intervals calculated of m2 obtained with \code{\link{predCap}} function.}
   \item{forecPlot3}{A plot of the forecasts values and the confidence intervals calculated of m3 obtained with \code{\link{predCap}} function.}
   \item{bandWidthComp}{A data frame with the forecasts band widths of each model in a column.}
   \item{bandWidthPlot}{A plot of the forecasts band widths of each model.}
}
\description{
Compare up to three ARIMA models of the class \code{\link{CArima}}, with theirs metrics (Sigma square, AIC and SBC) and their forecasting performance.
}
\examples{
#Some examples with AirPassengers
dates <- dateSeq(from = time(AirPassengers)[1], length.out = length(AirPassengers),
                 by = "month")
model1 <- CArima(as.vector(log(AirPassengers)), order = c(0, 1, 1), 
                 seasonal = list(order=c(0,0,1), period = 12))
model1
model2 <- CArima(as.vector(log(AirPassengers)), order = c(0, 1, 1), 
                 seasonal = list(order=c(0,1,1), period = 12))
model2 
model3 <- CArima(as.vector(log(AirPassengers)), order = c(1, 1, 0), 
                 seasonal = list(order=c(1, 1, 0), period = 12))
model3

comp <- compareARIMA(as.vector(log(AirPassengers)), dates, model1, model2, 
                     model3, h = 12, plot = F)
comp
}
\seealso{
\code{\link[ggplot2]{ggplot}}, \code{\link{predCap}}.
}
