% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formating.R
\name{dateSeq}
\alias{dateSeq}
\title{Generate Sequences of Dates for Time Series}
\usage{
dateSeq(from, to, length.out, by)
}
\arguments{
\item{from}{starting date.}

\item{to}{end date. Can be given instead of \code{length.out} parameter}

\item{by}{character string, increment of the sequence. One of "\code{min}", "\code{hour}", "\code{day}", "\code{week}", "\code{month}", "\code{quarter}", "\code{year}". See 'Details'}

\item{lenght.out}{desired length of the sequence. A non-negative integer. Can be given instead of \code{to} parameter.}
}
\value{
A vector of dates.
}
\description{
Creates a sequence of dates or times.
}
\details{
Depending on the value of \code{by}, there are some requirements for the other arguments.
\itemize{
   \item For "\code{min}" and "\code{hour}" the format of \code{from} and \code{to} should be a correct format for the function \code{\link[base]{as.POSIXct}}.
   \item For "\code{day}" and "\code{week}" the format of \code{from} and \code{to} should be a correct format for the function \code{\link[base]{as.Date}}.
   \item For "\code{month}" the format of \code{from} and \code{to} should be a correct format for the function \code{\link[zoo]{yearmon}}.
   \item For "\code{quarter}" the format of \code{from} and \code{to} should be a correct format for the function \code{\link[zoo]{yearqtr}}.
}
}
\examples{
dateSeq(from = "2022-01", length.out = 24, by = "month")
dateSeq(from = "2022-01-01", to = "2023-12", by = "month")
dateSeq(from = "2022-01-03", length.out = 12, by = "week")
dateSeq(from = "2022-01-03", to = "2022-03-21", by = "week")

#An example extracting the times of AirPassengers
dateSeq(from = time(AirPassengers)[1], length.out = length(AirPassengers), 
        by = "month")
}
\seealso{
\code{\link[base]{as.POSIXct}}, \code{\link[base]{as.Date}}, \code{\link[zoo]{yearmon}}, \code{\link[zoo]{yearqtr}}.
}
