% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_forecast.R
\name{predCap}
\alias{predCap}
\title{Forecasting Performance in ARIMA Models}
\usage{
predCap(object, h = 10, model, dates, plot = T)
}
\arguments{
\item{object}{a time series object or a matrix containing the time series values.}

\item{h}{number of points that will not be used to fit the model, but to use as testing. The points will be taken from the bottom.}

\item{model}{an object describing a time series model; e.g. of class \code{\link{CArima}}.}

\item{dates}{a vector or an array with the dates as \code{character} of each observation of the time series. This vector can be generated using the function \code{\link{dateSeq}}.}

\item{plot}{if \code{TRUE}, it will show all the plots in execution.}
}
\value{
A list containing at least the following components:
   \item{forecasts}{The forecasts of the \code{h} points extracted from the original data object, with their confidence intervals as a \code{data.frame}.}
   \item{SSEtable}{A \code{data.frame} containing the fitted values, real values, residual values and SSE values of the \code{h} points extracted from the original data object.}
   \item{forecPlot}{A plot of the forecasts values and the confidence intervals calculated.}
   \item{bandWidth}{The width of the forecasts confidence intervals.}
   \item{bandWidthPlot}{A plot of the forecasts confidence bands width.}
}
\description{
Performs more complete analysis of the forecasting performance of an ARIMA model by fitting the model with a train set with \code{h} less points and then use those \code{h} points as a test set.
}
\examples{
#Easy example with AirPassengers
dates <- dateSeq(from = time(AirPassengers)[1], length.out = length(AirPassengers),
                 by = "month")
model <- CArima(as.vector(log(AirPassengers)), order = c(0, 1, 1), 
                seasonal = list(order=c(0,1,1), period = 12))
model

predCapModel <- predCap(as.vector(log(AirPassengers)), h = 12, model, dates, 
                        plot = F) 
predCapModel
}
\seealso{
\code{\link[ggplot2]{ggplot}}.
}
