% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identification.R
\name{rmplot}
\alias{rmplot}
\title{Time Series Range-Mean Plot}
\usage{
rmplot(y, n = 5)
}
\arguments{
\item{y}{a vector or an array with the values of the time series.}

\item{n}{size of the subsamples to calculate the mean and range, this parameter should be equals to the period of the time series.}
}
\value{
A \code{\link[ggplot2]{ggplot}} object.
}
\description{
\code{rmplot} generates a ggplot object showing a range-mean plot of a time series.
}
\examples{
# Easy example with AirPassengers time series
rmplot(as.vector(AirPassengers), n = 12)
rmplot(log(as.vector(AirPassengers)), n = 12)
}
