% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identification.R
\name{tsplot}
\alias{tsplot}
\title{Time Series Plot}
\usage{
tsplot(x, y, n.unlabel = 9)
}
\arguments{
\item{x}{a vector or an array with the dates as \code{character} of each observation of the time series. This vector can be generated using the function \code{\link{dateSeq}}.}

\item{y}{a vector or an array with the values of the time series.}

\item{n.unlabel}{number of unlabeled dates between labeled dates in the plot.}
}
\value{
A \code{\link[ggplot2]{ggplot}} object.
}
\description{
\code{tsplot} creates a \code{\link[ggplot2]{ggplot}} object showing the main plot of a time series.
}
\examples{
# Easy example with AirPassengers time series
dates <- dateSeq(from = time(AirPassengers)[1], length.out = length(AirPassengers), 
                 by = "month")
tsplot(dates, as.vector(AirPassengers))
}
