% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identification.R
\name{tsspec}
\alias{tsspec}
\title{Time Series Frequency Spectrum}
\usage{
tsspec(y)
}
\arguments{
\item{y}{a vector or an array with the values of the time series.}
}
\value{
A list containing the following components: 
   \item{table}{A data frame with the values of the spectrum for each frequency represented.}
   \item{plot}{A \code{\link[ggplot2]{ggplot}} object of the frequency spectrum.}
}
\description{
\code{tsspec} generates a ggplot showing the frequency spectrum of a time series, and a table with the value of each frequency.
}
\examples{
# Easy example with AirPassengers time series
tsspec(as.vector(AirPassengers))
}
\seealso{
\code{\link[TSA]{periodogram}}
}
